Refactor cli operations
